refactor: separating modules to fix circular import
