+++ src/backend/langflow/utils/constants.py
@@ -1,7 +1,3 @@
-from langchain.agents import Tool
-from langflow.interface.custom_types import PythonFunction
-from langchain.agents.load_tools import get_all_tool_names
-
 OPENAI_MODELS = [
     "text-davinci-003",
     "text-davinci-002",
@@ -11,11 +7,8 @@ OPENAI_MODELS = [
 ]
 CHAT_OPENAI_MODELS = ["gpt-3.5-turbo", "gpt-4", "gpt-4-32k"]
 
-CUSTOM_TOOLS = {"Tool": Tool, "PythonFunction": PythonFunction}
 
 DEFAULT_PYTHON_FUNCTION = """
 def python_function(text: str) -> str:
     return text
 """
-
-ALL_TOOLS_NAMES = set(get_all_tool_names() + list(CUSTOM_TOOLS.keys()))
