+++ tests/test_endpoints.py
@@ -1,14 +1,4 @@
-# # build router
-# router = APIRouter()
-
-
-# @router.get("/all")
-# def get_all():
-#     return build_langchain_types_dict()
-
-
-# Buil test for /all endpoint
-from langflow.utils.constants import CUSTOM_TOOLS
+from langflow.interface.listing import CUSTOM_TOOLS
 from fastapi.testclient import TestClient
